% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_data}
\alias{age_data}
\title{age specimen data}
\format{
A data frame
\describe{
\item{loc}{long_lat}
\item{FMP_SUBAREA}{defined area for location}
}
}
\usage{
age_data
}
\description{
A dataset containing age info
}
\keyword{datasets}
