% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_error}
\alias{age_error}
\title{ageing error data}
\format{
A data frame
\describe{
\item{loc}{long_lat}
\item{FMP_SUBAREA}{defined area for location}
}
}
\usage{
age_error
}
\description{
A dataset containing age error info
}
\keyword{datasets}
