% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_banks.R
\docType{data}
\name{arrhythmic_item_bank}
\alias{arrhythmic_item_bank}
\title{Berkowitz Sight-Singing Item Bank filtered for SAA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}, \code{item_bank}) with 30921 rows and 44 columns.
}
\usage{
arrhythmic_item_bank
}
\description{
Berkowitz Sight-Singing Item Bank filtered for SAA
}
\keyword{datasets}
