% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-SAA.R
\name{aSAA}
\alias{aSAA}
\title{Deploy the adaptive SAA (aSAA)}
\usage{
aSAA(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  examples = 2,
  final_results = TRUE,
  musicassessr_aws = FALSE,
  store_results_in_db = FALSE,
  test_username = character(),
  gold_msi = TRUE,
  with_final_page = TRUE,
  item_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = character(),
  concise_wording = TRUE,
  skip_setup = FALSE
)
}
\arguments{
\item{skip_setup}{}
}
\value{

}
\description{
Deploy the adaptive SAA (aSAA)
}
