% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-SAA.R
\name{aSAA_standalone}
\alias{aSAA_standalone}
\title{Deploy adaptive SAA (aSAA) as standalone test}
\usage{
aSAA_standalone(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 15L, rhythmic = 10L),
  item_bank = Berkowitz::Berkowitz_IRT_arrhythmic_scaled,
  irt_model = Berkowitz::lm2.2_scaled,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = "",
  examples = 2L,
  final_results = TRUE,
  test_username = "",
  gold_msi = TRUE,
  with_final_page = TRUE,
  item_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = "",
  concise_wording = TRUE,
  skip_setup = FALSE,
  ...
)
}
\arguments{
\item{app_name}{Name of app.}

\item{num_items}{The number of items as a list.}

\item{item_bank}{The (IRT) item bank that will be used for selecting items.}

\item{irt_model}{The IRT model that will be used for selecting items.}

\item{demographics}{Deploy demographic form?}

\item{demo}{Is demo?}

\item{feedback}{Give feedback after trials?}

\item{admin_password}{psychTestR admin password.}

\item{SNR_test}{Deploy signal-to-noise ratio test?}

\item{get_range}{Deploy a test to get the users range at test time and present stimuli accordingly?}

\item{absolute_url}{If using online, absolute URL?}

\item{examples}{No of examples.}

\item{final_results}{Display final results?}

\item{test_username}{Is there a username for the user? This is different from a p_id.}

\item{gold_msi}{Deploy Gold-MSI form?}

\item{with_final_page}{Should there be a final page? FALSE if there will be more pages in the timeline.}

\item{item_length}{What melody lengths should the test be constrained to?}

\item{melody_sound}{Sound of melody? e.g, piano.}

\item{adjust_range}{Should the range of the user, recorded at test time, be adjusted based on heuristics?}

\item{test_name}{Custom name of the test.}

\item{show_socials}{Should social media sharing options be shown at the end?}

\item{headphones_test}{Should there be a headphone test?}

\item{get_user_info}{Grab user info via the browser?}

\item{microphone_test}{Deploy a microphone test?}

\item{allow_repeat_SNR_tests}{Logical. TRUE if participant can fail the SNR test threshold and try again.}

\item{append_trial_block_before}{A list of pages to go before the test.}

\item{append_trial_block_after}{A list of pages to go after the test.}

\item{stop_recording_after}{Stop recording after a certain amount of time.}

\item{max_goes}{How many goes can the user have per melody?}

\item{max_goes_forced}{Is this forced or optional?}

\item{long_tone_trials_as_screening}{Should long tone trials be used as a screening mechanism?}

\item{long_tone_trials_as_screening_failure_page}{Where should users be directed to if they fail the long tone screening?}

\item{success_on_completion_page}{Where should users be directed to when they complete successfully?}

\item{concise_wording}{TRUE for more detailed (but longer) instructions.}

\item{skip_setup}{TRUE to skip setup steps.}

\item{...}{}
}
\description{
Deploy adaptive SAA (aSAA) as standalone test
}
