% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz_item_bank_subset.R
\docType{data}
\name{Berkowitz_item_bank_subset}
\alias{Berkowitz_item_bank_subset}
\title{Pre-filtered Berkowitz Sight-Singing Item Bank (N-gram only)}
\format{
Dataframe.
}
\source{
\url{DOI:10.2307/3389405}
}
\usage{
Berkowitz_item_bank_subset
}
\description{
Pre-filtered Berkowitz Sight-Singing Item Bank (N-gram only)
}
\keyword{datasets}
