% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{len_bins}
\alias{len_bins}
\title{length bin data}
\format{
A data frame
\describe{
\item{loc}{long_lat}
\item{FMP_SUBAREA}{defined area for location}
}
}
\usage{
len_bins
}
\description{
A dataset containing length_bin info
}
\keyword{datasets}
