% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_data}
\alias{length_data}
\title{length specimen data}
\format{
A data frame
\describe{
\item{loc}{long_lat}
\item{FMP_SUBAREA}{defined area for location}
}
}
\usage{
length_data
}
\description{
A dataset containing length info
}
\keyword{datasets}
