% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST2.R
\name{MST2}
\alias{MST2}
\title{Deploy the MST2}
\usage{
MST2(
  aws_credentials,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  item_bank = itembankr::Berkowitz,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE
)
}
\arguments{
\item{get_range}{}
}
\value{

}
\description{
Deploy the MST2
}
