% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA.R
\name{SAA}
\alias{SAA}
\title{Deploy the SAA}
\usage{
SAA(
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  item_bank = Berkowitz::Berkowitz,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url,
  examples = 2,
  final_results = TRUE,
  musicassessr_state = "production",
  store_results_in_db = FALSE,
  test_username = character(),
  gold_msi = TRUE,
  with_final_page = TRUE,
  item_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  local_app_file_dir = "/Users/sebsilas/aws-musicassessr-local-file-upload/files/",
  sonic_annotator_local_location = "/Users/sebsilas/sonic-annotator",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  copy_audio_to_location = NULL,
  allow_repeat_SNR_tests = TRUE
)
}
\arguments{
\item{allow_repeat_SNR_tests}{}
}
\value{

}
\description{
Deploy the SAA
}
