% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA.R
\name{SAA_standalone}
\alias{SAA_standalone}
\title{Deploy SAA as standalone test}
\usage{
SAA_standalone(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  arrhythmic_item_bank = Berkowitz::ngram_item_bank,
  rhythmic_item_bank = Berkowitz::phrase_item_bank,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  examples = 2L,
  final_results = TRUE,
  musicassessr_aws = FALSE,
  store_results_in_db = FALSE,
  test_username = character(),
  gold_msi = TRUE,
  with_final_page = TRUE,
  melody_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = "https://adaptiveeartraining.com",
  concise_wording = TRUE,
  skip_setup = FALSE,
  full_screen = FALSE,
  validate_user_entry_into_test = FALSE,
  additional_scoring_measures = NULL,
  default_range = list(bottom_range = 48, top_range = 72),
  long_tone_paradigm = c("sing_along", "call_and_response"),
  ...
)
}
\arguments{
\item{app_name}{Name of app.}

\item{num_items}{The number of items as a list.}

\item{arrhythmic_item_bank}{The item bank (created with itembankr) to deployed with the test.}

\item{rhythmic_item_bank}{The item bank (created with itembankr) to deployed with the test.}

\item{demographics}{Deploy demographic form?}

\item{demo}{Is demo?}

\item{feedback}{Give feedback after trials?}

\item{admin_password}{psychTestR admin password.}

\item{SNR_test}{Deploy signal-to-noise ratio test?}

\item{get_range}{Deploy a test to get the users range at test time and present stimuli accordingly?}

\item{absolute_url}{If using online, absolute URL?}

\item{examples}{No of examples.}

\item{final_results}{Display final results?}

\item{musicassessr_aws}{Is this being deployed on AWS via the musicassessr setup?}

\item{store_results_in_db}{Store results in a database?}

\item{test_username}{Is there a username for the user? This is different from a p_id.}

\item{gold_msi}{Deploy Gold-MSI form?}

\item{with_final_page}{Should there be a final page? FALSE if there will be more pages in the timeline.}

\item{melody_length}{What melody lengths should the test be constrained to?}

\item{melody_sound}{Sound of melody? e.g, piano.}

\item{adjust_range}{Should the range of the user, recorded at test time, be adjusted based on heuristics?}

\item{test_name}{Custom name of the test.}

\item{show_socials}{Should social media sharing options be shown at the end?}

\item{headphones_test}{Should there be a headphone test?}

\item{get_user_info}{Grab user info via the browser?}

\item{microphone_test}{Deploy a microphone test?}

\item{allow_repeat_SNR_tests}{Logical. TRUE if participant can fail the SNR test threshold and try again.}

\item{append_trial_block_before}{A list of pages to go before the test.}

\item{append_trial_block_after}{A list of pages to go after the test.}

\item{stop_recording_after}{Stop recording after a certain amount of time.}

\item{max_goes}{How many goes can the user have per melody?}

\item{max_goes_forced}{Is this forced or optional?}

\item{long_tone_trials_as_screening}{Should long tone trials be used as a screening mechanism?}

\item{long_tone_trials_as_screening_failure_page}{Where should users be directed to if they fail the long tone screening?}

\item{success_on_completion_page}{Where should users be directed to when they complete successfully?}

\item{concise_wording}{TRUE for more detailed (but longer) instructions.}

\item{skip_setup}{TRUE to skip setup steps.}

\item{full_screen}{Should app be full screen?}

\item{validate_user_entry_into_test}{Should the user be validated against a session token?}

\item{additional_scoring_measures}{A function or list of functions with additional measures for scoring pYIN data.}

\item{default_range}{A list of the range that stimuli should be presented in, if not collected at test time.}

\item{long_tone_paradigm}{Can be sing_along or call_and_response.}

\item{...}{}
}
\value{

}
\description{
Deploy SAA as standalone test
}
