% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saa_waa.R
\name{saa_waa}
\alias{saa_waa}
\title{calculate size and weight at age}
\usage{
saa_waa(age_data, length_data, age_error, len_bins, rec_age)
}
\arguments{
\item{age_data}{goa trawl survey specimen data}

\item{length_data}{goa trawk survey length data}

\item{age_error}{age error matrix (since rockfish have a rectangle rather than square)}

\item{len_bins}{length bins used in assessment model}

\item{rec_age}{recruitment age used in assessment}
}
\description{
calculate size and weight at age
}
\examples{
\dontrun{
data("age_data")
data("length_data")
data("age_error")
saa_waa(age_data, length_data, age_error, len_bins=15:45, rec_age=2)
}

}
