% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST.R
\name{MST_standalone}
\alias{MST_standalone}
\title{Deploy MST as standalone test}
\usage{
MST_standalone(
  aws_credentials,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  item_bank = itembankr::Berkowitz,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url,
  examples = 2,
  final_results = TRUE,
  state = "production",
  store_results_in_db = FALSE,
  test_username = NULL,
  gold_msi = TRUE,
  item_length = c(3, 17)
)
}
\arguments{
\item{item_length}{}
}
\value{

}
\description{
Deploy MST as standalone test
}
