% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saa.R
\name{saa}
\alias{saa}
\title{size at age matrix}
\usage{
saa(age_data, length_data, len_bins, rec_age, age_error)
}
\arguments{
\item{age_data}{from ...}

\item{length_data}{from ...}

\item{len_bins}{length bins}

\item{rec_age}{recruitment start age of model}

\item{age_error}{ageing error matrix}
}
\description{
size at age matrix
}
