% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA.R
\name{SAA}
\alias{SAA}
\title{Deploy the SAA}
\usage{
SAA(
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  item_bank = itembankr::Berkowitz,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url,
  examples = 2,
  final_results = TRUE,
  state = "production",
  store_results_in_db = FALSE,
  test_username = NULL,
  gold_msi = TRUE,
  with_final_page = TRUE,
  item_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Melody Singing Task"
)
}
\arguments{
\item{test_name}{}
}
\value{

}
\description{
Deploy the SAA
}
