% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA.R
\name{SAA_standalone}
\alias{SAA_standalone}
\title{Deploy SAA as standalone test}
\usage{
SAA_standalone(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  num_examples = list(long_tones = 2L, arrhythmic = 2L, rhythmic = 0L),
  arrhythmic_item_bank = SAA::arrhythmic_item_bank,
  rhythmic_item_bank = SAA::rhythmic_item_bank,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  final_results = TRUE,
  test_username = character(),
  gold_msi = TRUE,
  with_final_page = TRUE,
  melody_length = c(4, 20),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = NULL,
  concise_wording = TRUE,
  skip_setup = FALSE,
  full_screen = FALSE,
  additional_scoring_measures = NULL,
  default_range = NULL,
  long_tone_paradigm = c("simultaneous_recall", "call_and_response"),
  get_p_id = FALSE,
  languages = c("en", "de", "it", "lv", "ch"),
  volume_meter_on_melody_trials = FALSE,
  volume_meter_on_melody_trials_type = "default",
  long_tone_length = 5,
  allow_SNR_failure = FALSE,
  requirements_page = TRUE,
  report_SNR = FALSE,
  show_introduction = TRUE,
  show_instructions = TRUE,
  asynchronous_api_mode = FALSE,
  experiment_id = NULL,
  user_id = NULL,
  get_answer_melodic = musicassessr::get_answer_pyin_melodic_production,
  content_border = "1px solid #e8e8e8",
  css = system.file("www/css/musicassessr.css", package = "musicassessr"),
  sample_item_bank_via_api = FALSE,
  pass_items_through_url_parameter = FALSE,
  show_intro_text = TRUE,
  show_microphone_type_page = TRUE,
  num_items_review = list(long_tones = 0L, arrhythmic = 0L, rhythmic = 0L),
  logo_url = NULL,
  dict = musicassessr::musicassessr_dict,
  redirect_on_failure_url = "https://google.com",
  sampler_function_arrhythmic = musicassessr::sample_arrhythmic,
  sampler_function_rhythmic = musicassessr::sample_rhythmic,
  use_presigned_url = TRUE,
  ...
)
}
\arguments{
\item{app_name}{Name of app.}

\item{num_items}{The number of items as a list.}

\item{num_examples}{The number of example items as a list.}

\item{arrhythmic_item_bank}{The item bank (created with itembankr) to deployed with the test.}

\item{rhythmic_item_bank}{The item bank (created with itembankr) to deployed with the test.}

\item{demographics}{Deploy demographic form?}

\item{demo}{Is demo?}

\item{feedback}{Give feedback after trials?}

\item{admin_password}{psychTestR admin password.}

\item{SNR_test}{Deploy signal-to-noise ratio test?}

\item{get_range}{Deploy a test to get the users range at test time and present stimuli accordingly?}

\item{absolute_url}{If using online, absolute URL?}

\item{final_results}{Display final results?}

\item{test_username}{Is there a username for the user? This is different from a p_id.}

\item{gold_msi}{Deploy Gold-MSI form?}

\item{with_final_page}{Should there be a final page? FALSE if there will be more pages in the timeline.}

\item{melody_length}{What melody lengths should the test be constrained to?}

\item{melody_sound}{Sound of melody? e.g, piano.}

\item{adjust_range}{Should the range of the user, recorded at test time, be adjusted based on heuristics?}

\item{test_name}{Custom name of the test.}

\item{show_socials}{Should social media sharing options be shown at the end?}

\item{headphones_test}{Should there be a headphone test?}

\item{get_user_info}{Grab user info via the browser?}

\item{microphone_test}{Deploy a microphone test?}

\item{allow_repeat_SNR_tests}{Logical. TRUE if participant can fail the SNR test threshold and try again.}

\item{append_trial_block_before}{A list of pages to go before the test.}

\item{append_trial_block_after}{A list of pages to go after the test.}

\item{stop_recording_after}{Stop recording after a certain amount of time.}

\item{max_goes}{How many goes can the user have per melody?}

\item{max_goes_forced}{Is this forced or optional?}

\item{long_tone_trials_as_screening}{Should long tone trials be used as a screening mechanism?}

\item{long_tone_trials_as_screening_failure_page}{Where should users be directed to if they fail the long tone screening?}

\item{success_on_completion_page}{Where should users be directed to when they complete successfully?}

\item{concise_wording}{TRUE for more detailed (but longer) instructions.}

\item{skip_setup}{TRUE to skip setup steps.}

\item{full_screen}{Should app be full screen?}

\item{additional_scoring_measures}{A function or list of functions with additional measures for scoring pYIN data.}

\item{default_range}{A list of the range that stimuli should be presented in, if not collected at test time.}

\item{long_tone_paradigm}{Can be simultaneous_recall or call_and_response.}

\item{get_p_id}{if TRUE, get the participant to enter their ID at the beginning of the test.}

\item{languages}{The available languages.}

\item{volume_meter_on_melody_trials}{Should there be a volume meter displayed on melody trial pages?}

\item{volume_meter_on_melody_trials_type}{If so, what type? Can be 'default' or 'playful'.}

\item{long_tone_length}{How long should the long tone be in seconds? Default: 5}

\item{allow_SNR_failure}{If TRUE, allow a participant/researcher to proceed, even if the SNR test is failed.}

\item{requirements_page}{Show a requirements page?}

\item{report_SNR}{Report SNR after test?}

\item{show_introduction}{Should introduction be shown (or skipped)?}

\item{show_instructions}{Should instructions be shown (or skipped)?}

\item{asynchronous_api_mode}{Should asynchronous API mode br shown?}

\item{experiment_id}{Manually give an experiment ID when using musicassessrdb.}

\item{user_id}{Manually give a user ID when using musicassessrdb.}

\item{get_answer_melodic}{The get_answer function for melodic trials.}

\item{content_border}{The psychTestR border style.}

\item{css}{Path to css stylesheet(s).}

\item{sample_item_bank_via_api}{Is the item bank being sampled via an API?}

\item{pass_items_through_url_parameter}{Are items being passed through a URL parameter?}

\item{show_intro_text}{Should intro text be shown?}

\item{show_microphone_type_page}{Should you ask the participant what kind of microphone they are using?}

\item{num_items_review}{Number of review items.}

\item{logo_url}{A URL for the psychTestR logo image.}

\item{dict}{Which dictionary to use for internationalisation.}

\item{redirect_on_failure_url}{A URL to redirect users to if an API check fails more than three times.}

\item{sampler_function_arrhythmic}{A psychTestR::code_block to determine how to sample from the arrhythmic item bank.}

\item{sampler_function_rhythmic}{A psychTestR::code_block to determine how to sample from the rhythmic item bank.}

\item{use_presigned_url}{For audio storing, should a presigned URL be used?}

\item{...}{}
}
\description{
Deploy SAA as standalone test
}
